+++ pages/index.js
@@ -40,48 +40,6 @@ export default function Home({ isConnected }) {
   
   const [habits, setHabits] = useState([]);
 
-  const handleSubmit = async (event) => {
-    // Stop the form from submitting and refreshing the page.
-    event.preventDefault()
-  
-    // Get data from the form.
-    const data = {
-      habit: event.target.habit.value,
-    }
-  
-    // Send the data to the server in JSON format.
-    const JSONdata = JSON.stringify(data)
-  
-    // API endpoint where we send form data.
-    const endpoint = '/api/form'
-  
-    // Form the request for sending data to the server.
-    const options = {
-      // The method is POST because we are sending data.
-      method: 'POST',
-      // Tell the server we're sending JSON.
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      // Body of the request is the JSON data we created above.
-      body: JSONdata,
-    }
-  
-    // Send the form data to our forms API on Vercel and get a response.
-    const response = await fetch(endpoint, options)
-  
-    // Get the response data from server as JSON.
-    // If server returns the name submitted, that means the form works.
-    const result = await response.json()
-  
-    const newHabit = {
-      key: uuid(),
-      text: result.data,
-    }
-    setHabits(current => [...current, newHabit]);
-    event.target.habit.value = "";
-  }
-
   const deleteHandler = (identifier) => {
     setHabits(habits.filter(habit => habit.key !== identifier));
   }
@@ -105,12 +63,12 @@ export default function Home({ isConnected }) {
         <title>Habit Tracker</title>
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <form onSubmit={handleSubmit} className={styles.form}>
+      {/* <form onSubmit={handleSubmit} className={styles.form}>
         <input className={styles.input} placeholder="Enter Habit" type="text" id="habit" name="habit" required />
         <button className={styles.submitButton} type="submit">Submit</button>
-      </form>
+      </form> */}
       <div className="grid">
-      <AddHabit />
+      <AddHabit editHandler={editHandler} habits={habits} setHabits={setHabits}/>
       {
         habits.map((habit, index) => {
           return <Habit key={habit.key} isEditable={true} color={cardColors[index % cardColors.length]} habit={habit} identifier={habit.key} deleteHandler={deleteHandler} editHandler={editHandler}/>
